+++ src/BursarExports.js
@@ -1,3 +1,50 @@
+import React, { useState } from 'react';
+import { useIntl } from 'react-intl';
+
+import {
+  Button,
+  Pane,
+  PaneFooter,
+} from '@folio/stripes/components';
+
+import { BursarExportsConfiguration } from './BursarExportsConfiguration';
+
 export const BursarExports = () => {
-  return 'BursarExports';
+  const { formatMessage } = useIntl();
+
+  const [bursarConfigForm, setBursarConfigForm] = useState();
+  const bursarConfigFormState = bursarConfigForm?.getState();
+
+  const saveBursarConfig = () => {
+    return bursarConfigForm?.submit();
+  };
+
+  const paneFooter = (
+    <PaneFooter
+      renderEnd={
+        <Button
+          buttonStyle="primary mega"
+          disabled={bursarConfigFormState?.pristine || bursarConfigFormState?.submitting}
+          onClick={saveBursarConfig}
+          type="submit"
+        >
+          {formatMessage({ id: 'ui-plugin-bursar-export.bursarExports.save' })}
+        </Button>
+      }
+    />
+  );
+
+  return (
+    <Pane
+      defaultWidth="fill"
+      footer={paneFooter}
+      id="pane-batch-group-configuration"
+      paneTitle={formatMessage({ id: 'ui-plugin-bursar-export.bursarExports' })}
+    >
+      <BursarExportsConfiguration
+        onFormStateChanged={setBursarConfigForm}
+        onSubmit={(values) => { console.log(values); }}
+      />
+    </Pane>
+  );
 };