+++ src/BursarExportsConfiguration/constants.js
@@ -1,0 +1,8 @@
+export const SCHEDULE_PERIODS = {
+  none: '',
+  hours: 'hours',
+  days: 'days',
+  weeks: 'weeks',
+};
+
+export const WEEKDAYS = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];