+++ src/BursarExportsConfiguration/validation.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { FormattedMessage } from 'react-intl';
+
+export function validateRequired(value) {
+  return value ? undefined : <FormattedMessage id="stripes-acq-components.validation.required" />;
+}