+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Stories
 * Project Setup: ui-plugin-bursar-export. Refs UIPBEX-1.
+* Bursar exports configuration form (UI). Refs UIPBEX-5.